refactor files names and folder structure
