+++ store/services/BaseService.ts
@@ -1,16 +1,10 @@
 import { createApi, fetchBaseQuery } from "@reduxjs/toolkit/query/react";
 
-const baseQuery = fetchBaseQuery({
-  baseUrl: `${process.env.API_BASE_URL}/`,
-  prepareHeaders: (headers) => {
-    headers.set("Content-Type", "application/json");
-    return headers;
-  },
-});
-
 export const BaseService = createApi({
   reducerPath: "baseService",
-  baseQuery: baseQuery,
+  baseQuery: fetchBaseQuery({
+    baseUrl: `${process.env.API_BASE_URL}/`,
+  }),
   tagTypes: [],
   endpoints: () => ({}),
 });
